eyes appearance bug fixed

+++ server/src/destroy.php
@@ -6,6 +6,8 @@ ini_set('display_errors', 1);
 $conn = dbConnect();
 $heroUUID = "";
 
+$destroyPrice = 1000;
+
 if ($conn->connect_error) {
     die("Database Connection Error: " . $conn->connect_error);
 }
@@ -38,7 +40,7 @@ if (isset($_GET['uuid'])) {
 
     if (validateUUID($heroUUID)) {
 
-        $sql = "SELECT * FROM entities WHERE uuid = '$uuid' AND type = 'eye'";
+        $sql = "SELECT * FROM entities WHERE private_uuid = '$heroUUID' AND type = 'hero'";
         $result = $conn->query($sql);
 
         if ($result->num_rows > 0) {
@@ -55,9 +57,20 @@ if (isset($_GET['uuid'])) {
                 exit(0);
             }  
                         
-            $entityBalance = $row['balance'];          
+            $heroBalance = $row['balance'];   
+            
+            if ($heroBalance < $destroyPrice) {
+                $response = array(
+                    'code' => 6,
+                    'message' => "Balance is not enough to destroy - $heroBalance/$destroyPrice",
+                    'entities' => []
+                );    
+                echo json_encode($response, JSON_UNESCAPED_UNICODE);
+                $conn->close();
+                exit(0);
+            }
 
-            $balanceUpdateSql = "UPDATE entities SET balance = balance + $entityBalance WHERE private_uuid = '$heroUUID'";
+            $balanceUpdateSql = "UPDATE entities SET balance = balance - $destroyPrice WHERE private_uuid = '$heroUUID'";
             $conn->query($balanceUpdateSql);
 
             $deleteSql = "DELETE FROM entities WHERE uuid = '$uuid'";
@@ -68,23 +81,29 @@ if (isset($_GET['uuid'])) {
                 'message' => "Entity $uuid caught!",
                 'entities' => []
             );
+            echo json_encode($response, JSON_UNESCAPED_UNICODE);
+            $conn->close();
+            exit(0);            
         } else {
             $response = array(
                 'code' => 1,
                 'message' => "Entity $uuid not foundd!",
                 'entities' => []
             );
+            echo json_encode($response, JSON_UNESCAPED_UNICODE);
+            $conn->close();
+            exit(0);            
         }
     } else {
         $response = array(
-            'code' => 4,
+            'code' => 5,
             'message' => "Invalid UUID format for $uuid",
             'entities' => []
         );
+        echo json_encode($response, JSON_UNESCAPED_UNICODE);
+        $conn->close();
+        exit(0);        
     }
-    echo json_encode($response, JSON_UNESCAPED_UNICODE);
-    $conn->close();
-    exit(0);
 } else {
     echo json_encode(array('code' => 3, 'message' => 'UUID parameter is missing'), JSON_UNESCAPED_UNICODE);
     $conn->close();
